Promote workflow module to directory
