+++ crates/core/tedge_api/src/workflow/script.rs
@@ -1,0 +1,53 @@
+use serde::de::Error;
+use serde::Deserialize;
+use serde::Deserializer;
+use serde::Serialize;
+use serde::Serializer;
+use std::fmt::Display;
+use std::fmt::Formatter;
+
+/// A parsed Unix command line
+#[derive(Clone, Debug, Eq, PartialEq)]
+pub struct ShellScript {
+    pub command: String,
+    pub args: Vec<String>,
+}
+
+/// Deserialize an Unix command line
+impl<'de> Deserialize<'de> for ShellScript {
+    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
+    where
+        D: Deserializer<'de>,
+    {
+        let command_line = String::deserialize(deserializer)?;
+        let mut args = shell_words::split(&command_line)
+            .map_err(|err| D::Error::custom(format!("invalid script: {command_line}: {err}")))?;
+        if args.is_empty() {
+            Err(D::Error::custom("invalid script: empty"))
+        } else {
+            let script = args.remove(0);
+            Ok(ShellScript {
+                command: script,
+                args,
+            })
+        }
+    }
+}
+
+/// Serialize an Unix command line, using appropriate quotes
+impl Serialize for ShellScript {
+    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
+    where
+        S: Serializer,
+    {
+        self.to_string().serialize(serializer)
+    }
+}
+
+impl Display for ShellScript {
+    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
+        let mut args = vec![self.command.clone()];
+        args.append(&mut self.args.clone());
+        f.write_str(&shell_words::join(args))
+    }
+}
