+++ crates/core/tedge_api/src/workflow/state.rs
@@ -1,366 +1,11 @@
-use crate::mqtt_topics::EntityTopicId;
-use crate::mqtt_topics::MqttSchema;
-use crate::mqtt_topics::OperationType;
-use log::info;
+use crate::workflow::WorkflowExecutionError;
 use mqtt_channel::Message;
-use mqtt_channel::QoS;
 use mqtt_channel::QoS::AtLeastOnce;
 use mqtt_channel::Topic;
-use serde::de::Error;
 use serde::Deserialize;
-use serde::Deserializer;
 use serde::Serialize;
-use serde::Serializer;
 use serde_json::json;
 use serde_json::Value;
-use std::collections::HashMap;
-use std::fmt::Display;
-use std::fmt::Formatter;
-
-pub type StateName = String;
-pub type OperationName = String;
-
-/// An OperationWorkflow defines the state machine that rules an operation
-#[derive(Clone, Debug, Serialize, Deserialize)]
-pub struct OperationWorkflow {
-    /// The operation to which this workflow applies
-    pub operation: OperationType,
-
-    /// Mark this workflow as built_in
-    #[serde(default, skip)]
-    pub built_in: bool,
-
-    /// The states of the state machine
-    #[serde(flatten)]
-    pub states: HashMap<StateName, OperationState>,
-}
-
-/// The current state of an operation request
-#[derive(Clone, Debug, Serialize, Deserialize)]
-pub struct OperationState {
-    /// Possibly a participant to which the action is delegated
-    pub owner: Option<String>,
-
-    /// Possibly a script to handle the operation when in that state
-    pub script: Option<ShellScript>,
-
-    /// Transitions
-    pub next: Vec<StateName>,
-}
-
-/// A parsed Unix command line
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub struct ShellScript {
-    pub command: String,
-    pub args: Vec<String>,
-}
-
-/// Deserialize an Unix command line
-impl<'de> Deserialize<'de> for ShellScript {
-    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-    where
-        D: Deserializer<'de>,
-    {
-        let command_line = String::deserialize(deserializer)?;
-        let mut args = shell_words::split(&command_line)
-            .map_err(|err| D::Error::custom(format!("invalid script: {command_line}: {err}")))?;
-        if args.is_empty() {
-            Err(D::Error::custom("invalid script: empty"))
-        } else {
-            let script = args.remove(0);
-            Ok(ShellScript {
-                command: script,
-                args,
-            })
-        }
-    }
-}
-
-/// Serialize an Unix command line, using appropriate quotes
-impl Serialize for ShellScript {
-    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: Serializer,
-    {
-        self.to_string().serialize(serializer)
-    }
-}
-
-impl Display for ShellScript {
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        let mut args = vec![self.command.clone()];
-        args.append(&mut self.args.clone());
-        f.write_str(&shell_words::join(args))
-    }
-}
-
-/// What needs to be done to advance an operation request in some state
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub enum OperationAction {
-    /// Nothing has to be done: simply move to the next step.
-    /// Such steps are intended to be overridden.
-    MoveTo(StateName),
-
-    /// The built-in behavior is used
-    BuiltIn,
-
-    /// The command is delegated to a participant identified by its name
-    Delegate(String),
-
-    /// Restart the device
-    Restart {
-        on_exec: StateName,
-        on_success: StateName,
-        on_error: StateName,
-    },
-
-    /// A script has to be executed
-    Script(ShellScript),
-
-    /// The command has been fully processed and needs to be cleared
-    Clear,
-}
-
-impl Display for OperationAction {
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        let str = match self {
-            OperationAction::MoveTo(step) => format!("move to {step} state"),
-            OperationAction::BuiltIn => "builtin".to_string(),
-            OperationAction::Delegate(owner) => {
-                format!("wait for {owner} to perform required actions")
-            }
-            OperationAction::Restart { .. } => "trigger device restart".to_string(),
-            OperationAction::Script(script) => script.to_string(),
-            OperationAction::Clear => "wait for the requester to finalize the command".to_string(),
-        };
-        f.write_str(&str)
-    }
-}
-
-/// Error preventing a workflow to be registered
-#[derive(thiserror::Error, Debug)]
-pub enum WorkflowDefinitionError {
-    #[error("Missing mandatory state: {state}")]
-    MissingState { state: StateName },
-
-    #[error("Missing transition for state: {state}")]
-    MissingTransition { state: StateName },
-}
-
-/// Error preventing a workflow to be registered
-#[derive(thiserror::Error, Debug)]
-pub enum WorkflowRegistrationError {
-    #[error("A workflow for this operation is already registered: {operation}")]
-    DuplicatedWorkflow { operation: OperationName },
-}
-
-/// Error preventing to infer the current action for an operation instance
-#[derive(thiserror::Error, Debug)]
-pub enum WorkflowExecutionError {
-    #[error("The command payload is not a JSON object")]
-    InvalidPayload(#[from] serde_json::Error),
-
-    #[error("Missing status in the command payload")]
-    MissingStatus,
-
-    #[error("No workflow is defined for the operation: {operation}")]
-    UnknownOperation { operation: OperationName },
-
-    #[error("No such step is defined for {operation}: {step}")]
-    UnknownStep {
-        operation: OperationName,
-        step: StateName,
-    },
-}
-
-/// Dispatch actions to operation participants
-#[derive(Default)]
-pub struct WorkflowSupervisor {
-    /// The user-defined operation workflow definitions
-    workflows: HashMap<OperationType, OperationWorkflow>,
-}
-
-impl WorkflowSupervisor {
-    /// Register a builtin workflow provided by thin-edge
-    pub fn register_builtin_workflow(
-        &mut self,
-        operation: OperationType,
-    ) -> Result<(), WorkflowRegistrationError> {
-        self.register_custom_workflow(OperationWorkflow::built_in(operation))
-    }
-
-    /// Register a user-defined workflow
-    pub fn register_custom_workflow(
-        &mut self,
-        workflow: OperationWorkflow,
-    ) -> Result<(), WorkflowRegistrationError> {
-        if let Some(previous) = self.workflows.get(&workflow.operation) {
-            if previous.built_in == workflow.built_in {
-                return Err(WorkflowRegistrationError::DuplicatedWorkflow {
-                    operation: workflow.operation.to_string(),
-                });
-            }
-
-            info!(
-                "The built-in {} operation has been customized",
-                workflow.operation
-            );
-            if workflow.built_in {
-                return Ok(());
-            }
-        }
-        self.workflows.insert(workflow.operation.clone(), workflow);
-        Ok(())
-    }
-
-    /// List the capabilities provided by the registered workflows
-    pub fn capability_messages(&self, schema: &MqttSchema, target: &EntityTopicId) -> Vec<Message> {
-        // To ease testing the capability messages are emitted in a deterministic order
-        let mut operations = self.workflows.values().collect::<Vec<_>>();
-        operations.sort_by(|&a, &b| a.operation.to_string().cmp(&b.operation.to_string()));
-        operations
-            .iter()
-            .map(|workflow| workflow.capability_message(schema, target))
-            .collect()
-    }
-
-    /// Extract the current action to be performed on a command request
-    ///
-    /// Returns:
-    /// - `Ok(Some(action)` when the request is well-formed
-    /// - `Ok(None)` when the request is finalized, i.e. when the command topic hase been cleared
-    /// - `Err(error)` when the request is ill-formed
-    pub fn get_workflow_current_action(
-        &self,
-        operation: &OperationType,
-        status: &Message,
-    ) -> Result<Option<(GenericCommandState, OperationAction)>, WorkflowExecutionError> {
-        self.workflows
-            .get(operation)
-            .ok_or_else(|| WorkflowExecutionError::UnknownOperation {
-                operation: operation.into(),
-            })
-            .and_then(|workflow| OperationWorkflow::get_operation_current_action(workflow, status))
-    }
-}
-
-impl OperationWorkflow {
-    /// Create a built-in operation workflow
-    pub fn built_in(operation: OperationType) -> Self {
-        let states = [
-            ("init", false, vec!["scheduled"]),
-            ("scheduled", true, vec!["executing"]),
-            ("executing", true, vec!["successful", "failed"]),
-            ("successful", false, vec![]),
-            ("failed", false, vec![]),
-        ]
-        .into_iter()
-        .map(|(step, delegate, next)| {
-            (
-                step.to_string(),
-                OperationState {
-                    owner: if delegate {
-                        Some("tedge".to_string())
-                    } else {
-                        None
-                    },
-                    script: None,
-                    next: next.into_iter().map(|s| s.to_string()).collect(),
-                },
-            )
-        })
-        .collect();
-
-        OperationWorkflow {
-            built_in: true,
-            operation,
-            states,
-        }
-    }
-
-    /// Return the MQTT message to register support for the operation described by this workflow
-    pub fn capability_message(&self, schema: &MqttSchema, target: &EntityTopicId) -> Message {
-        let meta_topic = schema.capability_topic_for(target, self.operation.clone());
-        let payload = "{}";
-        Message::new(&meta_topic, payload)
-            .with_retain()
-            .with_qos(QoS::AtLeastOnce)
-    }
-
-    /// Extract the current action to be performed on a command request
-    ///
-    /// Returns:
-    /// - `Ok(Some(action)` when the request is well-formed
-    /// - `Ok(None)` when the request is finalized, i.e. when the command topic hase been cleared
-    /// - `Err(error)` when the request is ill-formed
-    pub fn get_operation_current_action(
-        &self,
-        message: &Message,
-    ) -> Result<Option<(GenericCommandState, OperationAction)>, WorkflowExecutionError> {
-        match GenericCommandState::from_command_message(message) {
-            Ok(Some(cmd)) => self
-                .states
-                .get(&cmd.status)
-                .ok_or_else(|| WorkflowExecutionError::UnknownStep {
-                    operation: (&self.operation).into(),
-                    step: cmd.status.clone(),
-                })
-                .map(|state| {
-                    let action = OperationAction::from(state).inject_state(&cmd);
-                    Some((cmd, action))
-                }),
-            Ok(None) => Ok(None),
-            Err(err) => Err(err),
-        }
-    }
-}
-
-impl From<&OperationState> for OperationAction {
-    // TODO this must be called when an operation is registered, not when invoked.
-    fn from(state: &OperationState) -> Self {
-        match &state.script {
-            Some(script) if script.command == "restart" => {
-                let (on_exec, on_success, on_error) = match &state.next[..] {
-                    [] => ("executing", "successful", "failed"),
-                    [restarting] => (restarting.as_ref(), "successful", "failed"),
-                    [restarting, successful] => {
-                        (restarting.as_ref(), successful.as_ref(), "failed")
-                    }
-                    [restarting, successful, failed, ..] => {
-                        (restarting.as_ref(), successful.as_ref(), failed.as_str())
-                    }
-                };
-                OperationAction::Restart {
-                    on_exec: on_exec.to_string(),
-                    on_success: on_success.to_string(),
-                    on_error: on_error.to_string(),
-                }
-            }
-            Some(script) => OperationAction::Script(script.clone()),
-            None => match &state.owner {
-                Some(owner) if owner == "tedge" => OperationAction::BuiltIn,
-                Some(owner) => OperationAction::Delegate(owner.to_owned()),
-                None => match &state.next[..] {
-                    [] => OperationAction::Clear,
-                    [next] => OperationAction::MoveTo(next.to_owned()),
-                    _ => OperationAction::Delegate("unknown".to_string()),
-                },
-            },
-        }
-    }
-}
-
-impl OperationAction {
-    pub fn inject_state(self, state: &GenericCommandState) -> Self {
-        match self {
-            OperationAction::Script(script) => OperationAction::Script(ShellScript {
-                command: state.inject_parameter(&script.command),
-                args: state.inject_parameters(&script.args),
-            }),
-            _ => self,
-        }
-    }
-}
 
 /// Generic command state that can be used to manipulate any type of command payload.
 #[derive(Clone, Debug, Eq, PartialEq, Deserialize, Serialize)]
