+++ crates/core/tedge_api/src/workflow/supervisor.rs
@@ -1,0 +1,73 @@
+use crate::workflow::*;
+use log::info;
+
+/// Dispatch actions to operation participants
+#[derive(Default)]
+pub struct WorkflowSupervisor {
+    /// The user-defined operation workflow definitions
+    workflows: HashMap<OperationType, OperationWorkflow>,
+}
+
+impl WorkflowSupervisor {
+    /// Register a builtin workflow provided by thin-edge
+    pub fn register_builtin_workflow(
+        &mut self,
+        operation: OperationType,
+    ) -> Result<(), WorkflowRegistrationError> {
+        self.register_custom_workflow(OperationWorkflow::built_in(operation))
+    }
+
+    /// Register a user-defined workflow
+    pub fn register_custom_workflow(
+        &mut self,
+        workflow: OperationWorkflow,
+    ) -> Result<(), WorkflowRegistrationError> {
+        if let Some(previous) = self.workflows.get(&workflow.operation) {
+            if previous.built_in == workflow.built_in {
+                return Err(WorkflowRegistrationError::DuplicatedWorkflow {
+                    operation: workflow.operation.to_string(),
+                });
+            }
+
+            info!(
+                "The built-in {} operation has been customized",
+                workflow.operation
+            );
+            if workflow.built_in {
+                return Ok(());
+            }
+        }
+        self.workflows.insert(workflow.operation.clone(), workflow);
+        Ok(())
+    }
+
+    /// List the capabilities provided by the registered workflows
+    pub fn capability_messages(&self, schema: &MqttSchema, target: &EntityTopicId) -> Vec<Message> {
+        // To ease testing the capability messages are emitted in a deterministic order
+        let mut operations = self.workflows.values().collect::<Vec<_>>();
+        operations.sort_by(|&a, &b| a.operation.to_string().cmp(&b.operation.to_string()));
+        operations
+            .iter()
+            .map(|workflow| workflow.capability_message(schema, target))
+            .collect()
+    }
+
+    /// Extract the current action to be performed on a command request
+    ///
+    /// Returns:
+    /// - `Ok(Some(action)` when the request is well-formed
+    /// - `Ok(None)` when the request is finalized, i.e. when the command topic hase been cleared
+    /// - `Err(error)` when the request is ill-formed
+    pub fn get_workflow_current_action(
+        &self,
+        operation: &OperationType,
+        status: &Message,
+    ) -> Result<Option<(GenericCommandState, OperationAction)>, WorkflowExecutionError> {
+        self.workflows
+            .get(operation)
+            .ok_or_else(|| WorkflowExecutionError::UnknownOperation {
+                operation: operation.into(),
+            })
+            .and_then(|workflow| OperationWorkflow::get_operation_current_action(workflow, status))
+    }
+}
